Search: move passive Hero spotlight into its own library
Green origin-only e2ff784d36307b7eea46e40a893c362f37da2875 precedes move.
Exact approved e709 candidate plus one constructor-local no-key compatibility
ignore and narrowly approved shape inventory transfer. Same16inputs/0callbacks;
complete Spotlight/TitleArt/staticdeadURLmemo/geometry/feather/idleconstant.
Native code only2rect+6feather names; no engine/State/proxy/lifecycle changes.
No no-key API expansion. Five cross-library doc links become plain notation.

Body reconstruction: four ranges3-901,903-929,946-971,1776-1869 from origin047.
Exact after3identifier substitutions,5doclinks,1compatignore; LF checkout
normalization disclosed. Four origin SHA256 values:
63b99d57e18e22355684a7c8202348d516b97ae77dd0bd42563c1cb57bad7883
a8abf8cc230a8074038c808f6f22a7b079db9e42cf380067af0c8e7decf71431
853ae7cffef51c5b8a8e7837890d42663fd4245dd43cade55977f8fca6a21d76
6450b8e3add444342a7b89451e38957e9b5ecb8e883b83789b1dec063cd8cb0a
Local import/export/part graph101nodes no legacyhost/nativebackdrop path;
externalpackages not traversed. Not universal cycles/native proof.
Host6403->6404 +1; oldpart1993->947; newowner1060; whole+15.
ZERO host Leaves/638 credit; strict composition remains OPEN.

Verification pinnedFlutter3.44.8, each runtime only once as authorized:
flutter test --no-pub test/hero_spotlight_origin_test.dart
test/hero_presenter_origin_test.dart test/hero_presenter_runtime_test.dart
test/hero_trailer_backdrop_test.dart test/tv_hero_artwork_quality_test.dart
test/tv_home_stage_layouts_pin_test.dart test/theme/source_guard_test.dart
test/theme/shape_manifest_test.dart --reporter json
136PASS1knownFAIL0skip; exit1 JSONfalse,7.06s. Solefailure exact
no swept file grows a new bare radius: lib/widgets/tv_sidebar_nav.dart
already test/BASELINE_ALLOWLIST.txt7; source+allowlist unchanged. NOTallgreen.
Raw SHA30079a20f29e1e870e3984f2635da828f2c2cafb523cef370dc269c76cf1937d.

Shape transfer oldpart1->0/newradius-only1, total162unchanged, no swept claim.
Original parser/sidebar/favourite/token/floor490 loops unchanged.
Three actualfile probes targeted existing parser/membership each FAIL as intended:
oldradius0->1 SHA7b8fb533f68e67529b6019a2223e8f35ad4029dbb5a2ddd049025e666a10b187
newradius1->2 SHA67ad96efa39f795d2a594a8e8bd32891131d216b18ef4a13b77d218ec2a6bba7
removedentry SHA7bbcea80e294f32c26c76737a9106215520d004fd7b96e44abe697e852bd69e4
Exactcurrentbytes restored finally after each. Pristine3namedguardPASS
SHA72436ea2e2d4ddaf499336384af291acf04c486da7594f2e81b5dd8a41d1e92d.
Evidence-reader initially searched print instead of error JSON events; consumed
same firstprobe raw after repair, no runtime repetition.

Scoped5files:4inheritedcacheExtent infos, no new issues; initial5diagnostics
before authorized constructor-local ignore preserved.
py -3 tool/analyze_baseline.py: exit0,431current/449baseline,0new.
dart tool/check_layering.dart --json: exit0,56/77. Baseline unchanged.
These are currentchecker counts, not an independently rerun parentdelta.

Origin and all failed pre-move artifacts retained; pin01d50 unchanged.
Actualnativepositive/successful logo decode/listenerreplacement/allinterleavings
unproven. Windows/Linux public JSON-cache fixture boundary, no portability claim.
Future main union must preserve merged Sources9d4cf518/Atrium. No CODEMAP,
BOARD/NOTES, baseline, sourceguard or unrelated files modified.
Evidence .dart_tool/hero-prep/{applied-proof.json,production-eight-result.json,
shape-probes/summary.json,final-checks-result.json}; raw hashes above accessible
without local files. No selfmerge.
